Updated the inertia and convergence criterion
Using mean across sample for inertia so that average cosine distance is returned which makes inertia comparable across different sample sizes.

For the convergence criterion I normalize by both n_clusters and centroid_length  such that it is the mean squared error for each entry in the centroid this should make the convergence criterion more stable for varying numbers of centroids and centroid lengths.

However, existing code that does hyper-parameter sweeps may need updated.

Safest bet would be to ensure the centroids converge, but this hasn't been proven for the  current formulation